+++ src/main/java/k8s/example/client/Main.java
@@ -14,6 +14,7 @@ import org.slf4j.LoggerFactory;
 
 import io.kubernetes.client.openapi.models.V1Namespace;
 import io.kubernetes.client.openapi.models.V1NamespaceList;
+import k8s.example.client.audit.AuditController;
 import k8s.example.client.handler.UserDeleteJob;
 import k8s.example.client.k8s.K8sApiCaller;
 import k8s.example.client.metering.MeteringJob;
@@ -42,10 +43,13 @@ public class Main {
 			logger.info("[Main] Start Trial Namespace Timer");
 			startTrialNSTimer();
 			
+			// Start Audit
+			logger.info("[Main] Start Audit controller");
+			AuditController.start();
+			
 			// Start Start K8S watchers & Controllers
 			logger.info("[Main] Start K8S watchers");
 			K8sApiCaller.startWatcher(); // Infinite loop
-			
 			
 			
 		} catch (Exception e) {