+++ src/main/java/k8s/example/client/k8s/UserWatcher.java
@@ -49,12 +49,19 @@ public class UserWatcher extends Thread {
 					
 					// Logic here
 					try {
-						logger.info("[UserWatcher] Encrypt password of " + response.object.getUserInfo().getEmail());
+						logger.info("[UserWatcher] Encrypt password of User " + response.object.getMetadata().getName());
 
 						K8sApiCaller.encryptUserPassword(
-								response.object.getUserInfo().getEmail().replace("@", "-"), 
+								response.object.getMetadata().getName(), 
 								response.object.getUserInfo().getPassword(),
 								response.object);
+			    		
+			    		// Create UserSecurityPolicy with otpEnable false
+						try {
+							K8sApiCaller.createUserSecurityPolicy(response.object.getMetadata().getName());  
+							logger.info("[UserWatcher] UserSecurityPolicy of User" + response.object.getMetadata().getName() + " Create Success");
+						} catch (ApiException e) {
+						}
 					} catch (ApiException e) {
 						logger.info("ApiException: " + e.getMessage());
 						logger.info(e.getResponseBody());